+++ src/main/java/com/buuz135/simpleclaims/Main.java
@@ -10,11 +10,13 @@ import com.buuz135.simpleclaims.map.SimpleClaimsWorldMapProvider;
 import com.buuz135.simpleclaims.systems.events.*;
 import com.buuz135.simpleclaims.systems.tick.ChunkBordersTickingSystem;
 import com.buuz135.simpleclaims.systems.tick.EntryTickingSystem;
+import com.buuz135.simpleclaims.util.PartyInactivityThread;
 import com.buuz135.simpleclaims.systems.tick.TitleTickingSystem;
 
 import com.buuz135.simpleclaims.systems.tick.WorldMapUpdateTickingSystem;
 import com.hypixel.hytale.server.core.entity.entities.Player;
 import com.hypixel.hytale.server.core.event.events.player.AddPlayerToWorldEvent;
+import com.hypixel.hytale.server.core.event.events.player.PlayerDisconnectEvent;
 import com.hypixel.hytale.server.core.modules.interaction.interaction.config.Interaction;
 import com.hypixel.hytale.server.core.plugin.JavaPlugin;
 import com.hypixel.hytale.server.core.plugin.JavaPluginInit;
@@ -32,6 +34,8 @@ import java.util.logging.Level;
 public class Main extends JavaPlugin {
 
     public static Config<SimpleClaimsConfig> CONFIG;
+
+    private PartyInactivityThread partyInactivityTickingSystem;
 
     public Main(@NonNullDecl JavaPluginInit init) {
         super(init);
@@ -74,12 +78,19 @@ public class Main extends JavaPlugin {
         this.getEventRegistry().registerGlobal(AddPlayerToWorldEvent.class, (event) -> {
             var player = event.getHolder().getComponent(Player.getComponentType());
             var playerRef = event.getHolder().getComponent(PlayerRef.getComponentType());
-            ClaimManager.getInstance().setPlayerName(playerRef.getUuid(), player.getDisplayName());
+            ClaimManager.getInstance().setPlayerName(playerRef.getUuid(), player.getDisplayName(), System.currentTimeMillis());
+        });
+
+        this.getEventRegistry().registerGlobal(PlayerDisconnectEvent.class, (event) -> {
+            ClaimManager.getInstance().setPlayerName(event.getPlayerRef().getUuid(), event.getPlayerRef().getUsername(), System.currentTimeMillis());
         });
 
         var interaction = getCodecRegistry(Interaction.CODEC);
         interaction.register("UseBlock", ClaimUseBlockInteraction.class, ClaimUseBlockInteraction.CUSTOM_CODEC);
         interaction.register("CycleBlockGroup", ClaimCycleBlockGroupInteraction.class, ClaimCycleBlockGroupInteraction.CUSTOM_CODEC);
+
+        partyInactivityTickingSystem = new PartyInactivityThread();
+        partyInactivityTickingSystem.start();
     }
 
 }